+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 #### Added
 
-- ...
+- [#855](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/pull/855) Add helpers to create/change/drop a schema.
 
 ## v6.0.1